list got text with amount of pilots today, map got exclamation mark

+++ android/flywithme/src/main/java/net/exent/flywithme/service/ScheduleService.java
@@ -147,7 +147,7 @@ public class ScheduleService extends IntentService {
             con.setDoOutput(true);
             DataOutputStream outputStream = new DataOutputStream(con.getOutputStream());
             outputStream.writeByte(1);
-            outputStream.writeShort((int) takeoffId);
+            outputStream.writeShort(takeoffId);
             outputStream.writeLong(timestamp);
             outputStream.writeLong(pilotId);
             outputStream.writeUTF(name);
